Flag 5: clear remaining 57 int-conversion errors, promote -Werror=int-conversion
Final pass of the Category A int-conversion audit. Clears the last 57 errors
across 29 files and flips CMakeLists.txt:202 from -Wno-int-conversion to
-Werror=int-conversion.

Mechanical fixes (categories C/D/F from the handoff): (uintptr_t)&SYMBOL casts
for efParticleGetLoadBankID call sites; (TargetType *) casts on
(uintptr_t)base + (intptr_t)offset expressions passed to strongly-typed
parameters; (void *) wraps on SYVIDEO_DEFINE_FRAMEBUFFER_ADDR initializers.

LP64 landmines (category E):
- Widened gcFunc* helper family's u32 param -> uintptr_t param across
  objhelper.{c,h}, objscript.{c,h}, ifcommon.{c,h}, sc1pgame.{c,h}. Fixes
  a 64-bit pointer-truncation hazard in gcAddGObjScriptByLink where
  &gobjscript (stack pointer) was being passed through a u32.
- mnvsoptions.c: changed s32 sMNVSOptionsDamageGObj -> GObj * at the
  declaration; file-local symbol only used as a GObj pointer.
- ftmain.c:4886/4896/4913: added (intptr_t) on pointer-deref RHS stored into
  intptr_t event_file_head; replaced != NULL sentinel with != 0 to match.
- audio.c:1308: (s16 *)(uintptr_t) on PORT-override osVirtualToPhysical
  (returns u64 under PORT).
- Remaining sites (audio.c, dma.c, n_env.c, ef/efmanager.c, lb/*, mpcommon.c)
  are localised casts with justification recorded in the handoff doc.

All 158 int-conversion errors from the original audit are now resolved.
See docs/cmake_warning_audit_handoff_2026-04-20.md for per-site rationale
and follow-up concerns (dead s32 pointer fields in n_env.c FGM path,
func_ovl2_800EBC0C unused s32 arg).

+++ src/if/ifcommon.c
@@ -2745,7 +2745,7 @@ void ifCommonBattleInitPlacement(void)
 }
 
 // 0x80113638
-void ifCommonBattleInterfacePauseGObj(GObj *interface_gobj, u32 unused)
+void ifCommonBattleInterfacePauseGObj(GObj *interface_gobj, uintptr_t unused)
 {
     gcPauseGObjProcessAll(interface_gobj);
 
@@ -2753,7 +2753,7 @@ void ifCommonBattleInterfacePauseGObj(GObj *interface_gobj, u32 unused)
 }
 
 // 0x8011366C
-void ifCommonBattleInterfaceResumeGObj(GObj *interface_gobj, u32 unused)
+void ifCommonBattleInterfaceResumeGObj(GObj *interface_gobj, uintptr_t unused)
 {
     gcResumeGObjProcessAll(interface_gobj);
